refactor(mcp): address review — deduplicate resolveCollection, simplify loader injection
- Extract resolveCollection into helpers.ts (shared by query + trace)
- Move CollectionLoader type to helpers.ts to avoid circular deps
- Pass getCollection directly instead of redundant async wrapper
