+++ packages/mcp-server/src/tools/trace.ts
@@ -1,23 +1,8 @@
 import type { Embedder } from "@wtfoc/common";
 import { type TraceMode, trace } from "@wtfoc/search";
 import type { createStore } from "@wtfoc/store";
-import { loadCollection } from "../helpers.js";
-import type { CollectionLoader } from "../server.js";
-
-async function resolveCollection(
-	store: ReturnType<typeof createStore>,
-	collection: string,
-	collectionLoader?: CollectionLoader,
-) {
-	if (collectionLoader) {
-		const loaded = await collectionLoader(collection);
-		if (!loaded) throw new Error(`Collection "${collection}" not found`);
-		return loaded;
-	}
-	const head = await store.manifests.getHead(collection);
-	if (!head) throw new Error(`Collection "${collection}" not found`);
-	return loadCollection(store, head.manifest);
-}
+import type { CollectionLoader } from "../helpers.js";
+import { resolveCollection } from "../helpers.js";
 
 export async function handleTrace(
 	store: ReturnType<typeof createStore>,
